Refactor DataAnalysisTab into Chartview/DataAnalysisView and DataBalanceView
