+++ libs/dataset-explorer/src/index.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-export * from "./lib/DatasetExplorerTab";
+export * from "./lib/ChartView/DataAnalysisView/DatasetExplorerTab";
 export * from "./lib/DataAnalysisTab";
 export * from "./lib/TableView/TableView";
