+++ libs/dataset-explorer/src/lib/ChartView/DataAnalysisView/basePlotlyProperties.ts
