+++ libs/dataset-explorer/src/lib/ChartView/DataAnalysisView/buildScatterTemplate.ts
