+++ libs/dataset-explorer/src/lib/ChartView/DataAnalysisView/generatePlotlyProps.ts
