+++ libs/dataset-explorer/src/lib/ChartView/DataAnalysisView/getDatasetBar.ts
