+++ libs/dataset-explorer/src/lib/ChartView/DataAnalysisView/getDatasetBarOption.ts
