+++ libs/dataset-explorer/src/lib/ChartView/DataAnalysisView/getGroupedData.ts
