+++ libs/dataset-explorer/src/lib/DataAnalysisTab.tsx
@@ -13,8 +13,8 @@ import { localization } from "@responsible-ai/localization";
 import React from "react";
 
 import { dataAnalysisTabStyles } from "./DataAnalysisTab.styles";
-import { DataBalanceTab } from "./DataBalanceTab";
-import { DatasetExplorerTab } from "./DatasetExplorerTab";
+import { DataBalanceTab } from "./DataBalanceView/DataBalanceTab";
+import { DatasetExplorerTab } from "./ChartView/DataAnalysisView/DatasetExplorerTab";
 import { TableViewTab } from "./TableView/TableViewTab";
 
 interface IDataAnalysisTabProps {
